Bump cni-plugins to v1.3.0 and use Alpine 3.18.4
Build the binaries from source and add arguments for the base images.
This allows to build images for experimental architectures, like RISC-V,
given the right base images are provided as arguments.

Use cross-compilation instead of QEMU based compilation to speed up
multi-arch builds.

Build the final layer so that it solely contains the CNI plugin binaries
and Busybox to execute the installation script.